requirements.txt
black
flake8
ipython
jupyterlab
matplotlib
notebook
numpy
pandas
python-dotenv
scikit-learn
seaborn
catboost
xgboost
dill
#-e .
